# docs/api/openapi-specification.yaml
openapi: 3.0.3
info:
  title: オンライン診療システム API
  description: |
    ファストドクター主催「オンライン診療AI活用ハッカソン」のベースシステムAPI仕様書です。
    患者と医療従事者（医師・オペレータ・管理者）の両方に対応したRESTful APIを提供します。
  version: 1.0.0
  contact:
    name: ハッカソンサポートチーム
    email: hackathon@fastdoctor.jp
servers:
  - url: http://localhost:8787
    description: ローカル開発環境
  - url: https://api.online-medical.example.com
    description: 本番環境（Cloudflare Workers）
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
        details:
          type: object
          description: 詳細なエラー情報
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 総件数
        page:
          type: integer
          description: 現在のページ番号
        limit:
          type: integer
          description: 1ページあたりの件数
        hasNext:
          type: boolean
          description: 次のページが存在するか
        hasPrev:
          type: boolean
          description: 前のページが存在するか

    # ユーザー関連
    WorkerRole:
      type: string
      enum: [doctor, operator, admin]
      description: 医療従事者ロール
    Patient:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        address:
          type: string
        emergencyContact:
          type: string
        medicalHistory:
          type: array
          items:
            type: object
            properties:
              condition:
                type: string
              diagnosedDate:
                type: string
                format: date
              notes:
                type: string
        profileImageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Worker:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/WorkerRole'
        phoneNumber:
          type: string
        medicalLicenseNumber:
          type: string
        profileImageUrl:
          type: string
        isActive:
          type: boolean
        specialties:
          type: array
          items:
            $ref: '#/components/schemas/Specialty'
        qualifications:
          type: array
          items:
            $ref: '#/components/schemas/Qualification'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Specialty:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        isPrimary:
          type: boolean
          description: 主専門科かどうか（doctor_specialties経由）
    Qualification:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
        category:
          type: string
          enum: [specialist, certified, instructor, subspecialty, designated]
        certificateNumber:
          type: string
          description: 資格証番号（doctor_qualifications経由）
        acquiredDate:
          type: string
          format: date
          description: 取得日（doctor_qualifications経由）
        expiryDate:
          type: string
          format: date
          description: 有効期限（doctor_qualifications経由）

    # 予約関連
    AppointmentStatus:
      type: string
      enum: [scheduled, waiting, assigned, in_progress, completed, cancelled]
      description: 予約ステータス
    AppointmentType:
      type: string
      enum: [initial, follow_up, emergency]
      description: 診察種別
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patient:
          $ref: '#/components/schemas/Patient'
        workerId:
          type: integer
        worker:
          $ref: '#/components/schemas/Worker'
        scheduledAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        chiefComplaint:
          type: string
          description: 主訴
        meetingId:
          type: string
          description: Chime Meeting ID
        appointmentType:
          $ref: '#/components/schemas/AppointmentType'
        durationMinutes:
          type: integer
          default: 30
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # 問診票関連
...
  responses:
...
    ForbiddenError:
      description: 権限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: FORBIDDEN
            message: このリソースへのアクセス権限がありません
    NotFoundError:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: 指定されたリソースが見つかりません
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: VALIDATION_ERROR
            message: 入力値が不正です
            details:
              field: email
              reason: 有効なメールアドレスではありません
security:
  - BearerAuth: []
paths:
...
  /api/patient/appointments:
    get:
      tags:
        - Patient
      summary: 予約一覧取得
      description: 自分の予約一覧を取得します
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
          description: ステータスでフィルタ
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Appointment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Patient
      summary: 新規予約作成
      description: 新しい診察予約を作成します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scheduledAt
                - chiefComplaint
                - appointmentType
              properties:
                scheduledAt:
                  type: string
                  format: date-time
                chiefComplaint:
                  type: string
                  description: 主訴
                appointmentType:
                  $ref: '#/components/schemas/AppointmentType'
                durationMinutes:
                  type: integer
                  default: 30
      responses:
        '201':
          description: 予約作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/patient/appointments/{id}:
    get:
      tags:
        - Patient
      summary: 予約詳細取得
      description: 指定した予約の詳細情報を取得します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Patient
      summary: 予約変更
      description: 予約日時や主訴を変更します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduledAt:
                  type: string
                  format: date-time
                chiefComplaint:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Patient
      summary: 予約キャンセル
      description: 予約をキャンセルします
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: キャンセル成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 予約をキャンセルしました
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/worker/patients:
    get:
      tags:
        - Worker
      summary: 患者一覧取得
      description: 患者一覧を取得します
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: 名前・メールアドレス・電話番号で検索
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/worker/patients/{id}:
    get:
      tags:
        - Worker
      summary: 患者詳細取得
      description: 指定した患者の詳細情報を取得します
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Worker
      summary: 患者情報更新
      description: 患者情報を更新します（管理者のみ）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phoneNumber:
                  type: string
                address:
                  type: string
                emergencyContact:
                  type: string
                medicalHistory:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/worker/appointments:
    get:
      tags:
        - Worker
      summary: 全予約一覧
      description: 全ての予約一覧を取得します
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
        - name: workerId
          in: query
          schema:
            type: integer
          description: 担当医師でフィルタ
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: 予約日でフィルタ
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Appointment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...